requirements.txt
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
Django==3.0.2
django-cors-headers==3.2.1
djangorestframework==3.11.0
idna==2.9
pkg-resources==0.0.0
pytz==2019.3
requests==2.23.0
sqlparse==0.3.0
stripe==2.43.0
urllib3==1.25.8
